requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-wtf>=1.2.0
authlib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
